backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.5
pydantic[email]==2.12.5
email-validator==2.3.0
PyJWT==2.11.0
bcrypt==4.1.3
python-dotenv==1.2.1
openai==1.99.9
razorpay==2.0.0
reportlab==4.4.10
python-multipart==0.0.22
